[python] Move the None-operand rejection out of the binop dispatcher
get_binary_operator_expr sits at CCN 227 against a gate of 15, so any decision
point added there fails the complexity check. Put the rejection and the
delegation in handle_none_operand instead: the dispatcher loses the branch it
gained, handle_none_comparison is untouched, and the new function is CCN 2.

Behaviour is unchanged.
